feat(site): move about -> homepage, homepage -> about
