+++ apps/main-site/src/app/(main-site)/_components/GoogleResult.tsx
