+++ apps/main-site/src/app/(main-site)/_components/QuestionAnswerArea.tsx
